Fix subtel bugs in InsertionSorter
Fix some indexing bugs.
When the array is of size 1 adding 2 to first is indexing
out of range. When the size is more than that it will be
useless to start from the second item since the first
is the minimal but I decided that ifs will make the
code more complicated.